Исправлены команды; переименнованы классы сервисов; сделано оповещение пользователя о событиях приложения; Удалены все ненужные абстракции и расширения коллекций;

+++ Services/DataStorage.cs
@@ -1,18 +1,18 @@
 ﻿using MemeFolder.Domain.Models;
-using MemeFolder.Domain.Models.AbstractModels;
 using MemeFolder.Extentions;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace MemeFolder.Services
 {
     public delegate void SearchServiceEvent(IEnumerable<Meme> folderObjects);
-    public delegate void FolderEvent(Folder folder);
-    public delegate void MemeEvent(Meme meme);
+    public delegate void FolderEvent(Folder folder, string message);
+    public delegate void MemeEvent(Meme meme, string message);
+    public delegate void MemeTagEvent(MemeTag memeTag, string message);
 
     public class DataStorage
     {
@@ -20,6 +20,8 @@ namespace MemeFolder.Services
         private readonly IMemeDataService _memeDataService;
         private readonly IMemeTagDataService _memeTagDataService;
         private readonly IMemeTagNodeDataService _memeTagNodeDataService;
+
+        private readonly ClientConfigService _clientConfigService;
         
 
         public ObservableCollection<Folder> Folders { get; }
@@ -38,6 +40,11 @@ namespace MemeFolder.Services
         public event MemeEvent OnEditMeme;
         public event MemeEvent OnRemoveMeme;
 
+        public event MemeTagEvent OnAddMemeTag;
+        public event MemeTagEvent OnEditMemeTag;
+        public event MemeTagEvent OnRemoveMemeTag;
+
+
         public void NavigateByRequest(Func<Meme,bool> func)
         {
             IEnumerable<Meme> memes = Memes.Where(func);
@@ -50,16 +57,70 @@ namespace MemeFolder.Services
             NewRequest?.Invoke(memes);
         }
 
+        private string GetFolderAnotherName(string rootPath, string title)
+        {
+            string newFolderPath = string.Empty;
+            int num = 1;
+            while (true)
+            {
+                string tempTitle = $"{title} ({num++})";
+
+                newFolderPath = @$"{rootPath}\{tempTitle}";
+                if (!Directory.Exists(newFolderPath))
+                {
+                    Directory.CreateDirectory(newFolderPath);
+                    break;
+                }
+            }
+
+            return newFolderPath;
+        }
+
+        private string GetMemeAnotherName(string rootPath, string title, string imagePath)
+        {
+            string newMemePath = string.Empty;
+            int num = 1;
+            while (true)
+            {
+                newMemePath = @$"{rootPath}\{title} ({num++}){Path.GetExtension(imagePath)}";
+                if (!File.Exists(newMemePath))
+                {
+                    File.Copy(imagePath, newMemePath);
+                    break;
+                }
+            }
+
+            return newMemePath;
+        }
 
         public async Task<Folder> AddFolder(Folder folder, Folder parentFolder)
         {
+            string newFolderPath = string.Empty;
+            if (string.IsNullOrEmpty(folder.Title))
+            {
+                newFolderPath = GetFolderAnotherName(parentFolder.FolderPath, "Новая папка");
+                folder.Title = Path.GetDirectoryName(newFolderPath);
+                folder.FolderPath = newFolderPath;
+            }
+            else
+            {
+                newFolderPath = @$"{parentFolder.FolderPath}\{folder.Title}";
+                if (Directory.Exists(newFolderPath))
+                {
+                    newFolderPath = GetFolderAnotherName(parentFolder.FolderPath, folder.Title);
+                }
+
+                Directory.CreateDirectory(newFolderPath);
+                folder.FolderPath = newFolderPath;
+            }
+            
             Folder createdFolderEntity = await _folderDataService.Create(folder);
             if (createdFolderEntity != null)
             {
                 createdFolderEntity.ParentFolder = parentFolder;
                 parentFolder.Folders.Add(createdFolderEntity);
                 Folders.Add(createdFolderEntity);
-                OnAddFolder?.Invoke(createdFolderEntity);
+                OnAddFolder?.Invoke(createdFolderEntity, "Папка успешно добавлена");
                 
                 return createdFolderEntity;
             }
@@ -72,7 +133,7 @@ namespace MemeFolder.Services
             if (updatedFolderEnitiy != null)
             {
                 updatedFolderEnitiy.OnAllPropertyChanged();
-                OnEditFolder?.Invoke(updatedFolderEnitiy);
+                OnEditFolder?.Invoke(updatedFolderEnitiy, "Папка успешно изменена");
             }
         }
 
@@ -90,31 +151,70 @@ namespace MemeFolder.Services
                     folderFR.Memes.ToList().ForEach(m =>
                     {
                         Memes.Remove(m);
-                        OnRemoveMeme?.Invoke(m);
+                        OnRemoveMeme?.Invoke(m, null);
                     });
                     Folders.Remove(folderFR);
-                    OnRemoveFolder?.Invoke(folderFR);
+                    OnRemoveFolder?.Invoke(folderFR, null);
                 });
-               
-                OnRemoveFolder?.Invoke(folder);
+
+                GC.Collect();
+                GC.WaitForPendingFinalizers();
+
+                OnRemoveFolder?.Invoke(folder, "Папка успешно удалена");
             }
         }
 
         public async Task AddMeme(Meme meme, Folder folder)
         {
-            meme.ImageData = MemeExtentions.ConvertImageToByteArray(meme.ImagePath);
-            Meme createdMemeEntity = await _memeDataService.Create(meme);
-            if (createdMemeEntity != null)
+            try
             {
-                createdMemeEntity.Image = MemeExtentions.ConvertByteArrayToImage(createdMemeEntity.ImageData);
-                createdMemeEntity.Image.Freeze();
-                createdMemeEntity.ImageData = null;
-                folder.Memes.Add(createdMemeEntity);
-                folder.OnPropertyChanged(nameof(folder.Memes));
-                Memes.Add(createdMemeEntity);
+                string newMemePath = @$"{folder.FolderPath}\{meme.Title}{Path.GetExtension(meme.ImagePath)}";
+                if (File.Exists(newMemePath))
+                {
+                    newMemePath = GetMemeAnotherName(folder.FolderPath, meme.Title, meme.ImagePath);
+                    meme.Title = Path.GetFileNameWithoutExtension(newMemePath);
+                    meme.ImagePath = newMemePath;
+                }
+                else
+                {
+                    File.Copy(meme.ImagePath, newMemePath);
+                    meme.ImagePath = newMemePath;
+                }
 
-                OnAddMeme?.Invoke(createdMemeEntity);
+                meme.ImageData = MemeExtentions.ConvertImageToByteArray(meme.ImagePath);
+
+                List<MemeTagNode> memeTagNodes = new List<MemeTagNode>();
+                foreach(MemeTagNode mtn in meme.Tags.ToArray())
+                {
+                    memeTagNodes.Add(mtn);
+                    meme.Tags.Remove(mtn);
+                }
+
+                Meme createdMemeEntity = await _memeDataService.Create(meme);
+                if (createdMemeEntity != null)
+                {
+                    createdMemeEntity.Image = MemeExtentions.ConvertByteArrayToImage(createdMemeEntity.ImageData);
+                    createdMemeEntity.Image.Freeze();
+                    createdMemeEntity.ImageData = null;
+                    folder.Memes.Add(createdMemeEntity);
+                    folder.OnPropertyChanged(nameof(folder.Memes));
+                    Memes.Add(createdMemeEntity);
+
+                    memeTagNodes.ForEach(async (newMemeTagNode) =>
+                    {
+                        newMemeTagNode.Meme = createdMemeEntity;
+                        MemeTagNode dbCreatedMemeTagNode = await _memeTagNodeDataService.Create(newMemeTagNode);
+                        createdMemeEntity.Tags.Add(dbCreatedMemeTagNode);
+                    });
+
+                    OnAddMeme?.Invoke(createdMemeEntity, "Мем успешно добавлен");
+                }
             }
+            catch (Exception ex)
+            {
+
+            }
+            
         }
 
         public async Task EditMeme(Meme meme)
@@ -172,7 +272,7 @@ namespace MemeFolder.Services
                 updatedMemeEnitiy.ImageData = null;
                 updatedMemeEnitiy.OnAllPropertyChanged();
 
-                OnEditMeme?.Invoke(updatedMemeEnitiy);
+                OnEditMeme?.Invoke(updatedMemeEnitiy, "Мем успешно изменён");
             }
         }
 
@@ -184,7 +284,7 @@ namespace MemeFolder.Services
                 folder.Memes.Remove(meme);
                 Memes.Remove(meme);
 
-                OnRemoveMeme?.Invoke(meme);
+                OnRemoveMeme?.Invoke(meme, "Мем успешно удалён");
             }
         }
 
@@ -194,6 +294,8 @@ namespace MemeFolder.Services
             if (createdMemeTagEnitiy != null)
             {
                 MemeTags.Add(createdMemeTagEnitiy);
+
+                OnAddMemeTag?.Invoke(createdMemeTagEnitiy, "Тег успешно добавлен");
             }
         }
 
@@ -203,28 +305,33 @@ namespace MemeFolder.Services
             if (updatedMemeTagEnitiy != null)
             {
                 updatedMemeTagEnitiy.OnAllPropertyChanged();
+                
+                OnEditMemeTag?.Invoke(updatedMemeTagEnitiy, "Тег успешно изменён");
             }
         }
 
         public async Task RemoveMemeTag(MemeTag memeTag)
         {
-            IEnumerable<Meme> memes = Memes.Where(meme => meme.Tags.FirstOrDefault(m => m.MemeTag.Id == memeTag.Id) != null);
-            foreach (Meme meme in memes)
+            if (await _memeTagDataService.Delete(memeTag.Id))
             {
-                MemeTagNode memeTagNode = meme.Tags.FirstOrDefault(mtn => mtn.MemeTag.Id == memeTag.Id);
-                meme.Tags.Remove(memeTagNode);
-                await _memeTagNodeDataService.Delete(memeTagNode.Id);
-            }
-            MemeTags.Remove(memeTag);
+                IEnumerable<Meme> memes = Memes.Where(meme => meme.Tags.FirstOrDefault(m => m.MemeTag.Id == memeTag.Id) != null);
+                foreach (Meme meme in memes)
+                {
+                    MemeTagNode memeTagNode = meme.Tags.FirstOrDefault(mtn => mtn.MemeTag.Id == memeTag.Id);
+                    meme.Tags.Remove(memeTagNode);
+                    await _memeTagNodeDataService.Delete(memeTagNode.Id);
+                }
+                MemeTags.Remove(memeTag);
 
-            await _memeTagDataService.Delete(memeTag.Id);
+                OnRemoveMemeTag?.Invoke(memeTag, "Тег успешно удалён");
+            }      
         }
 
-        public DataStorage(IMemeDataService memeDataService,
+        public DataStorage(Guid rootGuid,
+            IMemeDataService memeDataService,
             IFolderDataService folderDataService,
             IMemeTagDataService memeTagDataService,
             IMemeTagNodeDataService memeTagNodeDataService)
-           
         {
             _memeDataService = memeDataService;
             _folderDataService = folderDataService;
@@ -234,6 +341,7 @@ namespace MemeFolder.Services
             Folders = new ObservableCollection<Folder>();
             Memes = new ObservableCollection<Meme>();
             MemeTags = new ObservableCollection<MemeTag>();
+
             foreach (MemeTag memeTag in _memeTagDataService.GetAll().Result)
                 MemeTags.Add(memeTag);
 
@@ -259,13 +367,11 @@ namespace MemeFolder.Services
                     }
             }
 
-
             Folders.Add(folder);
 
             foreach (Folder innerFolder in folder.Folders)
                 GetDataRecursively(innerFolder);
         }
 
-       
     }
 }
